requirements.txt
allure-pytest==2.14.3
selenium==4.31.00
pytest==8.3.5
pytest-xdist==3.6.1

h11~=0.14.0
pip~=21.2.4
attrs~=25.3.0
wheel~=0.37.0
PySocks~=1.7.1
trio~=0.29.0
outcome~=1.3.0.post0
sniffio~=1.3.1
exceptiongroup~=1.2.2
sortedcontainers~=2.4.0
idna~=3.10
pluggy~=1.5.0
iniconfig~=2.1.0
certifi~=2025.1.31
urllib3~=2.4.0
trio-websocket~=0.12.2
websocket-client~=1.8.0
packaging~=24.2
wsproto~=1.2.0
allure-python-commons~=2.14.3